Cloud Functions/gcs-to-bq/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==3.3.1
google-cloud==0.34.0
google-cloud-storage==2.6.0
google-api-python-client==2.68.0